NSInteger isn't defined in the tiger frameworks, which we use in our includes for dep-debug. Doesn't break [21372]

+++ Frameworks/AIUtilities Framework/Source/IKPictureTakerForTiger.h
@@ -104,5 +104,5 @@
 @end
 
 @interface NSObject (IKPictureTakerForTigerDelegate)
-- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
+- (void)pictureTakerDidEnd:(id)inPictureTaker returnCode:(int)returnCode contextInfo:(void *)contextInfo;
 @end
